% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_api.R
\name{album_search}
\alias{album_search}
\title{Search for a specific album}
\usage{
album_search(name = NULL, similar_to = NULL, decade = NULL,
  genre = NULL, country = NULL, gender = NULL, limit = NULL,
  offset = NULL, query = NULL, top_rated = NULL, api_key = NULL)
}
\arguments{
\item{name}{*optional Return albums who match the name; exact, case insensitive     "White+album"}

\item{similar_to}{*optional Return albums who are similar to the given name; exact, case insensitive similar_to="The wall"}

\item{decade}{*optional a decade for instance}

\item{genre}{*optional a genre such as Blues, Rock etc.}

\item{country}{see the dictionary}

\item{gender}{male or female}

\item{limit}{*optional max number of results defaults to 20}

\item{offset}{*optional pagination which page of the results do you want}

\item{query}{use this if you know exactly what your query is, leave NULL if you
have want to build up the query in steps}

\item{top_rated}{only the top results}

\item{api_key}{your developer API key see \link{getting_api_key}}
}
\value{
a response object.
}
\description{
Search for a specific album, the name needs to match exactly. But you can
also search for similar albums but the name needs to match exactly too.
See for genres and decade valid options:
\url{https://developer.musicgraph.com/api-docs/v2/dictionary}
}
\seealso{
Other album_endpoint: \code{\link{album_edges}},
  \code{\link{album_lookup}}
}
