% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists_api.R
\name{artist_edges}
\alias{artist_edges}
\title{Artist Edges}
\usage{
artist_edges(artistID, fields = NULL, api_key = NULL)
}
\arguments{
\item{artistID}{an artistID or artistID/similar, artistID/albums or artistID/tracks}

\item{fields}{*optional f.i. name, id, spotify_id, gender, popularity etc.}

\item{api_key}{your developer API key see \link{getting_api_key}}
}
\value{
a response object.
}
\description{
Search for more information about artists. Either general information or
similar artists, albums or tracks.
}
\details{
either artist id or id/similar, id/albums id/tracks
possibly fields
}
\seealso{
Other artists_endpoint: \code{\link{artist_lookup}},
  \code{\link{artist_search}}
}
