% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists_api.R
\name{artist_search}
\alias{artist_search}
\title{Search for an artist}
\usage{
artist_search(name = NULL, similar_to = NULL, decade = NULL,
  genre = NULL, country = NULL, gender = NULL, limit = NULL,
  offset = NULL, query = NULL, api_key = NULL)
}
\arguments{
\item{name}{*optional Return artists who match the name; exact, case insensitive     "John+Lennon"}

\item{similar_to}{*optional Return artists who are similar to the given name; exact, case insensitive     &similar_to=Pink+Floyd}

\item{decade}{*optional a decade for instance}

\item{genre}{*optional a genre such as Blues, Rock etc.}

\item{country}{see the dictionary for more info about the country}

\item{gender}{male or female}

\item{limit}{*optional max number of results defaults to 20}

\item{offset}{*optional pagination which page of the results do you want}

\item{query}{use this if you know exactly what your query is, leave NULL if you
have want to build up the query in steps}

\item{api_key}{your developer API key see \link{getting_api_key}}
}
\value{
a response object.
}
\description{
Search for a specific artist, the name needs to match exactly. But you can
also search for similar artists but the name needs to match exactly too.
}
\details{
\url{https://developer.musicgraph.com/api-docs/v2/dictionary}
}
\seealso{
Other artists_endpoint: \code{\link{artist_edges}},
  \code{\link{artist_lookup}}
}
