% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{result_to_dataframe}
\alias{result_to_dataframe}
\title{Take the result from a query and move it to a data.frame}
\usage{
result_to_dataframe(object)
}
\arguments{
\item{object}{an object created with f.i. artist_lookup().}
}
\value{
a data.frame
}
\description{
This function takes an object and translates it into a data.frame.
}
