% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_api.R
\name{track_edges}
\alias{track_edges}
\title{Album Edges}
\usage{
track_edges(albumID, fields = NULL, api_key = NULL)
}
\arguments{
\item{albumID}{an albumID or albumID/artists or albumID/tracks}

\item{fields}{*optional f.i. name, id, spotify_id, gender, popularity etc.}

\item{api_key}{your developer API key see \url{getting_api_key}}
}
\value{
a response object.
}
\description{
Search for more information about a album Either general information or
artists of the album or tracks.
}
\details{
possibly fields
}
\seealso{
Other tracks_endpoint: \code{\link{track_lookup}},
  \code{\link{track_search}}
}
