% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{validate_artistID}
\alias{validate_artistID}
\title{Check artistID against pattern}
\usage{
validate_artistID(artistID)
}
\arguments{
\item{artistID}{an artistID}
}
\description{
When it doesn't match the pattern a warning will be given but the function
does continue.
}
\seealso{
Other validators: \code{\link{validate_decade}},
  \code{\link{validate_gender}},
  \code{\link{validate_genre}}
}
