% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{validate_decade}
\alias{validate_decade}
\title{Internal function to check for decade}
\usage{
validate_decade(decadestring)
}
\arguments{
\item{decadestring}{a decade such as 1980s}
}
\value{
correctly formatted decade that matches the list on musicgraph.com
}
\description{
This function matches decade against a list of possible decades.
It will throw an error when it doesn't match
}
\seealso{
Other validators: \code{\link{validate_artistID}},
  \code{\link{validate_gender}},
  \code{\link{validate_genre}}
}
