% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{validate_gender}
\alias{validate_gender}
\title{Internal function to check for gender}
\usage{
validate_gender(gender)
}
\arguments{
\item{gender}{male or female}
}
\description{
This function checks if the gender is male or female.
It will throw an error when it doesn't match
}
\seealso{
Other validators: \code{\link{validate_artistID}},
  \code{\link{validate_decade}},
  \code{\link{validate_genre}}
}
