% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_interface.R
\name{validate_genre}
\alias{validate_genre}
\title{Internal function to validate genre}
\usage{
validate_genre(genrestring)
}
\arguments{
\item{genrestring}{the genre string people give in searches}
}
\value{
a string formatted in the correct way that matches the genrelist.
}
\description{
It changes the string to titlecase if not already so
and checks it against the list found on the website.
The function fails / returns an error (I hope informative enough) when it
can't find a match.
I made this list for v2 of the API on 2017-02-16. It might
change. I hope not.
\url{https://developer.musicgraph.com/api-docs/v2/dictionary}
}
\seealso{
Other validators: \code{\link{validate_artistID}},
  \code{\link{validate_decade}},
  \code{\link{validate_gender}}
}
